Refactoring Spectrum Analysis to a Separate File
